Updated modal logic and renamed employee id view
