--- front/src/views/PixelFeedbackView.vue
@@ -1,5 +1,0 @@
-<template>
-</template>
-
-<script setup lang="ts">
-</script>
